Don't mane "reasonable" move in search-perf
This was just picking the first move from the move generator, which
isn't actually that reasonable taken out far enough -- it was leading to
a mate after 11 moves or so, due to dumb moves! Use the search at every
ply, but reset the zobrist to force a re-search every time for better
perf measurements.