Implement magic bitboards
This seems to take `perft -d 6` from 8.2s to 7.0s and `search-perf 9 11`
from 38.2s to 32.4s, quite the speedup!

Descriptions of how magic bitboards work are all over the internet, at
http://pradu.us/old/Nov27_2008/Buzz/research/magic/Bitboards.pdf for
example. This particular implementation uses constants taken from
http://pradu.us/old/Nov27_2008/Buzz/releases/magicmoves.zip -- there is
probably room to improve this from the tech in the decade since then,
but this is good for now. (Only the constants were take, no other code.)