Swap Undo to State
An idea from Stockfish. The "undo" need not be a copy, but can be the
canonical way we store things. Saves some copying on undoing a move, but
more importantly lays the groundwork for being able to keep cached some
very expensive things we may want in the future.

For some reason this is like 2% slower in `test-perft`? Still the
refactor is important for bigger speedups.

Does not yet implement "statelist" for "permanent" moves, instead just
leaking memory. Also removes `stress` which I don't think is useful any
more.