Compute full king_attackers mask
So we can see if we are in check vs double-check easily. This is a perf
loss right now, probably due to unconditionally computing the mask for
the to-move player even when the pseudolegal move did not turn out to be
legal (when, previously, we'd only compute for the off-move player). As
the generator is updated to generate fewer illegal pseudomoves, this
will be less of an issue. I hope.