Move to "fancy" bitboards
I tried this a few times and it didn't work out, but this time it seems
to help by about 1% or a tiny bit less in `search-perf 8 8`. I think
what helped is the `Entry` struct, which gives better data locality than
just looking up in the various tables?

This is a perf loss for perft though. I suspect that the reduced cache
pressure keeps more of the transposition table in cache?