Return board index functions into macros
This change seems to be responsible for the perf loss (especially
`board_row_of`). This looks to be a quirk of the clang optimizer -- see
https://godbolt.org/z/7nzra5d4x for example; those two `if` blocks
should compile to the same assembly but they do not. In the particular
case here, with the surrounding code, this causes some extra register
pressure and spills to the stack, which causes the perf issues.