Add game to investigate later
The current version, playing as "new", lost this game. Something about
the repetition detection and the tt, perhaps, seems to have caused the
engine to not see a problem coming -- see its response in particular
after e5c6. The aspiration failures make it *seem* like it knows
something might be wrong but it doesn't find it. A very weird trace.