Move timer to clock_gettime-based
Instead of using signals. On the one hand, it's a little sad to use
polling when we could be directly told that time is up. On the other
hand, it avoid some UNIX-y wrangling with signals and volatile etc. But
most important, it makes it straightforward to manipulate this end time,
e.g., to take more time in event of an aspiration failure or less if we
aren't going to get another ply.