Wire SEE into moveiter
Config option to order captures by SEE. As expected, this seems to be a
wash right now -- the more expensive SEE offsets the win of a smaller
search tree. This will hopefully get better with some potential
improvements to SEE as well as with using it to prune losing captures in
qsearch.

Disabled for now pending further testing and improvements.